chore(firstMile): move CreateToken out of steps/python into steps/
